Add HTTP NOT FOUND code when resource doesn't exist

+++ src/main/java/com/github/mauronunez/course/restapi/controller/CommonController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.github.mauronunez.course.restapi.service.CommonService;
 
@@ -26,7 +27,12 @@ public class CommonController<K,T> {
 	}
 
 	@GetMapping("/{id}") public T byId(@PathVariable("id") K id) {
-		return commonService.byId(id);
+		T instance= commonService.byId(id);
+		if(instance==null) {
+			throw new ResponseStatusException(
+			          HttpStatus.NOT_FOUND, "Instance Not Found");
+		}
+		return instance;
 	}
 	
 	@PostMapping public ResponseEntity<Void> create(@RequestBody T course) {
@@ -35,10 +41,20 @@ public class CommonController<K,T> {
 	}
 	
 	@PutMapping("/{id}") public void update(@PathVariable("id") K id, @RequestBody T course) {
+		T instance= commonService.byId(id);
+		if(instance==null) {
+			throw new ResponseStatusException(
+			          HttpStatus.NOT_FOUND, "Instance Not Found");
+		}
 		commonService.update(id, course);
 	}
 
 	@DeleteMapping("/{id}") public void delete(@PathVariable("id") K id) {
+		T instance= commonService.byId(id);
+		if(instance==null) {
+			throw new ResponseStatusException(
+			          HttpStatus.NOT_FOUND, "Instance Not Found");
+		}
 		commonService.delete(id);
 	}
